+++ src/main/java/com/bank/api/customer/pojo/LoginInputValue.java
@@ -6,9 +6,9 @@ import lombok.Data;
 @Data
 public class LoginInputValue {
 	private String card_number;
-	private int code;
+	private double code;
 
-	public LoginInputValue(String card_number, int code) {
+	public LoginInputValue(String card_number, double code) {
 		this.card_number = card_number;
 		this.code = code;
 	}
